lane-agentic-core/python-agentic-core/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.7.4,<3.0.0
httpx==0.27.0
python-dotenv==1.0.1

# AI / Agents
google-generativeai==0.8.3
langchain==0.3.0
langchain-google-genai==2.0.7
langgraph==0.2.22
huggingface_hub==0.25.0

# Audio Processing
google-cloud-speech==2.26.0
speechrecognition==3.10.4
Pillow==10.4.0

# Optional Audio processing if needed (RawNet2 stub config doesn't run full pytorch to save space locally, we will use lightweight client or API)
# torch==2.0.0
# torchaudio==2.0.0
